Logout user on http 401 code
