+++ ios-app/UI/ProfileViewController.swift
@@ -131,27 +131,7 @@ class ProfileViewController: UIViewController {
             style: UIAlertActionStyle.destructive,
             handler: { action in
                 
-                let fcmToken = UserDefaults.standard.string(forKey: Constants.FCM_TOKEN)
-                let deviceToken = UserDefaults.standard.string(forKey: Constants.DEVICE_TOKEN)
-                
-                if (fcmToken != nil && deviceToken != nil ) {
-                    let parameters: Parameters = [
-                        "device_id": deviceToken!,
-                        "registration_id": fcmToken!,
-                        "platform": "ios"
-                    ]
-                    
-                    TPApiClient.apiCall(endpointProvider: TPEndpointProvider(.unRegisterDevice), parameters: parameters,
-                                        completion: { _, _ in})
-                }
-                UIApplication.shared.unregisterForRemoteNotifications()
-
-                // Clear only user related tables
-                DBInstance.clearTables()
-                TPApiClient.apiCall(endpointProvider: TPEndpointProvider(.logout), completion: {_,_ in})
-                // Logout on Facebook
-                LoginManager().logOut()
-                KeychainTokenItem.clearKeychainItems()
+                UIUtils.logout()
                 let loginViewController = self.storyboard?.instantiateViewController(withIdentifier:
                     Constants.LOGIN_VIEW_CONTROLLER) as! LoginViewController
